Rename src/ to lib/, since its not fair for them to be in src/ while `main.go` is not.
Also, rename the package to `lib` to follow the general convention to keep package name same as the folder.
